requirements.txt
tqdm
ipdb
pyyaml
wandb==0.15.0
flax
tensorflow-cpu==2.9.0
tensorflow_datasets
tensorflow_io==0.26.0
tensorflow_hub
moviepy
protobuf==3.20.3
lpips-jax
ffmpeg-python
internetarchive
cloudpickle
crafter
jaxlib
numpy
optax
rich
ruamel.yaml
tensorflow_probability
dm_control
opencv-python
gym==0.17.3